+++ NEWS.md
@@ -2,6 +2,7 @@
 * Options added to ft_dyns to return the modulus of the 2D spectra and to
 pre-transform the chemical shift axis if needed.
 * Bug fix for printing lists of mrs_data objects.
+* comb_coils function now suports lists of mrs_data objects.
 
 # spant 1.19.0
 * Added sum_mrs_list function.